requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
google-generativeai
joblib
requests
